Mapping Updated!
Changed the way mapping works so that each tile is not a new instance of
a tile object, but just a copy. This should improve performance. Also
the map should look a little different, and I fixed a bug or two related
to where the player character appears and is handled upon instantiation.